refactor(otelcol/filter): move non-arg types to own file
