+++ component/otelcol/processor/filter/filter.go
@@ -26,19 +26,6 @@ func init() {
 	})
 }
 
-type traceConfig struct {
-	Span      []string `river:"span,attr,optional"`
-	SpanEvent []string `river:"spanevent,attr,optional"`
-}
-
-type metricConfig struct {
-	Metric    []string `river:"metric,attr,optional"`
-	Datapoint []string `river:"datapoint,attr,optional"`
-}
-type logConfig struct {
-	LogRecord []string `river:"log_record,attr,optional"`
-}
-
 type Arguments struct {
 	// ErrorMode determines how the processor reacts to errors that occur while processing a statement.
 	ErrorMode ottl.ErrorMode `river:"error_mode,attr,optional"`
